requirements.txt
a2a-sdk[http-server]
uvicorn
python-dotenv
langchain
langchain-openai
